\name{B2mIntra_98}
\docType{data}
\alias{B2mIntra_98}
\title{Example Data B2mIntra_98.VFP (Beta-2-microglobulin RIA) from the Variance Function Program 12.0 from Sadler}
\description{
  Results in this file are all serum/CSF clinical specimen duplicates during 1998 from a low throughput Beta-2-microglobulin RIA.

}
\usage{data(B2mIntra_98)}
\format{data.frame with 554 observations and 2 variables.}
\references{
       www.aacb.asn.au/AACB/Resources/Variance-Function-Program
}
\keyword{datasets}

 

