% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{.onLoad}
\alias{.onLoad}
\title{Keep backward compatibility by assigning new to old function names when
loading the package. Old name did not comply to new CRAN package check rules.}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{(character) string giving the library directory where
the package defining the namespace was found}

\item{pkgname}{(character) string giving the name of the package}
}
\description{
Keep backward compatibility by assigning new to old function names when
loading the package. Old name did not comply to new CRAN package check rules.
}
