% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mon}
\alias{get_mon}
\title{Determine Monotony of Vector.}
\usage{
get_mon(x)
}
\arguments{
\item{x}{(numeric) vector}
}
\value{
1 if monotonically increasing, -1 if monotonically decreasing, otherwise,
        a vector of 1 or -1 indicating increasing or decreasing status
}
\description{
Determine Monotony of Vector.
}
\examples{
x1 <- seq(-1, 1, .1)
x2 <- seq(1, -1, -.1)
y <- x1^2 / 2 + x1/3 - 5
get_mon(x1)
get_mon(x2)
get_mon(y)
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}
