\name{MultiLotReproResults}
\docType{data}
\alias{MultiLotReproResults}
\title{Result of a Real-World Precsion Experiment on 10 Samples}
\description{
 (data.frame) representing the results of an imprecision experiment.
 There are 10 observations and four variables which can be used directly
 as input for function 'fit.vfp'. Different samples are indicated by 
 rownames.
}
\usage{MultiLotReproResults}
\format{data.frame with 10 observations and 4 variables.}
\keyword{datasets}

 

