% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfp.R
\name{powfun4}
\alias{powfun4}
\title{Internal Function Model 4.}
\usage{
powfun4(x, potenz)
}
\arguments{
\item{x}{(numeric) parameter 1}

\item{potenz}{(numeric) parameter 2}
}
\description{
Internal Function Model 4.
}
