% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{powfun4simple}
\alias{powfun4simple}
\title{Internal Function Model 4.}
\usage{
powfun4simple(x, potenz)
}
\arguments{
\item{x}{(numeric) parameter 1}

\item{potenz}{(numeric) parameter 2}
}
\description{
Internal Function Model 4.
}
