% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfp.R
\name{powfun5simple}
\alias{powfun5simple}
\title{Internal Function Model 5.}
\usage{
powfun5simple(x, K)
}
\arguments{
\item{x}{(numeric) parameter 1}

\item{K}{(numeric) parameter 2}
}
\description{
Internal Function Model 5.
}
