% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfp.R
\name{powfun8}
\alias{powfun8}
\title{Internal Function Model 8.}
\usage{
powfun8(x, C, signJ)
}
\arguments{
\item{x}{(numeric) parameter 1}

\item{C}{(numeric) parameter 2}

\item{signJ}{(numeric) parameter 3}
}
\description{
Internal Function Model 8.
}
