% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfp.R
\name{predict.modelEP17}
\alias{predict.modelEP17}
\title{Predict Method for Objects of Class 'modelEP17'.}
\usage{
\method{predict}{modelEP17}(object, newdata = NULL, alpha = 0.05,
  ci.type = c("vc", "sd", "cv"), CI.method = c("chisq", "t", "normal"),
  use.log = FALSE, ...)
}
\arguments{
\item{object}{(object) of class 'modelEP17'}

\item{newdata}{(numeric) vector of data points at which prediction shall be made}

\item{alpha}{(numeric) value defining the 100(1-alpha)\% confidence interval for
predictions}

\item{ci.type}{(character) string specifying on which scale prediction shall be made}

\item{CI.method}{(character) string specifying which type of CI shall be used}

\item{use.log}{(logical) TRUE X- and Y-values will be returned on log-scale}

\item{...}{additional arguments}
}
\description{
This is a helper function not intented to be used directly.
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}
