\name{RealData1}
\docType{data}
\alias{RealData1}
\title{Result of a Real-World Precsion Experiment on 10 Samples}
\description{
 (data.frame) representing the results of an imprecision experiment.
 There are 10 observations and six variables which can be used directly
 as input for function 'fit.vfp'. Different samples are indicated by variable
 "No".
}
\usage{RealData1}
\format{data.frame with 10 observations and 6 variables.}
\keyword{datasets}

 

