\name{ReproData}
\docType{data}
\alias{ReproData}
\title{Serum Work Area Example Data of a Reproducibility Experiment}
\description{
  Multi-site precision experiment using 6 human serum-samples and 2 precision
  control samples, which were all measured at each site on 12 days, with 2 runs
  per day and 2 replicates per run, i.e. there are 8 (sample) x 3 (sites/labs) x
  12 (days) x 2 (runs) x 2 (replicates) = 1152 observations overall.
}
\usage{data(ReproData)}
\format{data.frame with 1152 observations and 5 variables.}

\keyword{datasets}

 

