% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfp.R
\name{Signif}
\alias{Signif}
\title{Adapted Version of Function 'signif'}
\usage{
Signif(x, digits = 4, force = TRUE, ...)
}
\arguments{
\item{x}{(numeric) value to be rounded to the desired number
of significant digits}

\item{digits}{(integer) number of significant digits}

\item{force}{(logical) TRUE = force the return value to have at least 4 significant
digits, i.e. to integers with less digits zeros will be appended after
the decimal separator, otherwise the return value will be casted from
character to numeric}

\item{...}{additional parameters}
}
\value{
number with 'digits' significant digits, if 'force=TRUE' "character" objects will be
			returned otherwise objects of mode "numeric"
}
\description{
This function adapts base-function \code{\link{signif}}
by always returning integer values in case the number of
requested significant digits is less than the the number of
digits in front of the decimal separator.
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}
