\name{T4S9_99}
\docType{data}
\alias{T4S9_99}
\title{Example Data T4S9_99.VFP (T4 RIA) from the Variance Function Program 12.0 from Sadler}
\description{
  Clinical specimen T4 RIA duplicates produced during 1999 by a staff member.
}
\usage{T4S9_99}
\format{data.frame with 8553 observations and 2 variables.}
\references{
       VFP Program Version 12.0\\
       \href{http://www.aacb.asn.au/professionaldevelopment/useful-tools}{http://www.aacb.asn.au/professionaldevelopment/useful-tools}
}
\keyword{datasets}

 

