\name{Glucose}
\docType{data}
\alias{Glucose}
\title{Example Data Glucose.VFP from the Variance Function Program 12.0 from Sadler}
\description{
  Intra-day precision of arterial glucose measurement on the Precision PCx instrument 
  (Abbott Laboratories, Abbott Park, Illinois, USA).  
  The data were kindly provided by Peter Watkinson, The John Radcliffe Hospital, Oxford, UK.  
  They provide an excellent example of a situation where a simple variance function 
  should be imposed in preference to accepting the fit of one of the flexible functions.

}
\usage{data(Glucose)}
\format{data.frame with 206 observations and 2 variables.}
\references{
       www.aacb.asn.au/AACB/Resources/Variance-Function-Program
}
\keyword{datasets}

 

