\name{T4S9_99}
\docType{data}
\alias{T4S9_99}
\title{Example Data T4S9_99.VFP (T4 RIA) from the Variance Function Program 12.0 from Sadler}
\description{
  Clinical specimen T4 RIA duplicates produced during 1999 by a staff member.
}
\usage{data(T4S9_99)}
\format{data.frame with 8553 observations and 2 variables.}
\references{
       www.aacb.asn.au/AACB/Resources/Variance-Function-Program
}
\keyword{datasets}

 

